requirements.txt
Django==2.1.4
django-bootstrap4==0.0.7
django-crispy-forms==1.7.2
django-simple-search==1.0.2
Pillow==5.4.1
pytz==2018.7
validate-email==1.3
